requirements.txt
numpy==1.24.2
pandas==1.5.3
tqdm==4.64.1
matplotlib==3.6.3
scikit-learn-1.2.1
toml==0.10.2
torch-summary==1.4.5

# other installation
# pip install torch==1.8.0+cu111 torchvision==0.9.0+cu111 -f https://download.pytorch.org/whl/torch_stable.html